Вынес CheckPieceMove в отдельный класс
